requirements.txt
autokeras
dotenv
foolbox
numpy
pydantic
pytorch_lightning
requests
tensorflow
torch
torchvision
transformers
